Update Form.vue
addContact function: 
Its better pass a object to the array, with the properties name, phone and e-mail.

Show contacts: 
Loop through the Array with objects and show the properties on the object.

Delete a contact:
To delete a specific contact, first we need to get the Index position of this object in the array. To do that, we have to modify the for loop, and add the index option. The index option will store the index of each object in the array.

Now that we have the index, we need to pass this index to the deleteList function.

deleteList function:
It receives the index of that object and use the splice method (from the array) to take the object out from the list.

The splice method receive the Index of the object, and the number of itens that you want to erase from the list. In our case is just one
splice(index, 1);